Raise exceptions in send_bulk_message after loop has run
Catch when a device registration id fails on the API and raise an exception if none of the device ids pass and a different exception if some of the ids pass. Either way, good ids are sent, but the program will raise an exception to say that not all ids have been set.