Update client.py
First of all thank you so much for this project, it's brilliant and it helped me out a lot! There is just one minor correction that I made. The "extra" key needs to be inside the "aps" key for it to work. so I changed it to aps_data.update(extra) instead of data.update(extra) and moved it to run before data["aps"] = aps_data.
Let me know what you think and thanks again!